# src/main/java/com/example/backend/controller/response/UserResponse.java
package com.example.backend.controller.response;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;
import java.util.UUID;
@Data
@NoArgsConstructor
@AllArgsConstructor
public class UserResponse {
    private UUID id;
    private String username;
    private String role;
    private List<InventoryBookResponse> inventories;
    private String token;
}